Updates README.
Apparently, problem input isn't supposed to be shared. I disagree with the logic there but I have to respect the request and licensing of AoC.

A real coder would scrub the git history, but a real competition would surface their request not to share input better. So it's a wash.